railway-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
mediapipe>=0.10.11
opencv-python-headless==4.10.0.84
numpy==1.26.4
httpx==0.27.2
yt-dlp==2024.9.27
supabase==2.8.1
python-multipart==0.0.9
python-dotenv==1.0.1
scenedetect[opencv]>=0.6.4
